Factor out transaction helpers and CSV export to util.js and add jsdoc comments
- factored transaction helpers / CSV export into static w/ light instance wrappers
- moved static (most of code) to util.js so can be `require`-ed separately
  (for browser-laptop #4332)
- added jsdoc comments
